campuscompanion/api/pyproject.toml
[tool]
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = [ "Marcus Schiesser <mail@marcusschiesser.de>" ]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
llama-index = "^0.9.19"
pypdf = "^3.17.0"
python-dotenv = "^1.0.0"
docx2txt = "^0.8"
torch = "^2.2.1"
sentence-transformers = "^2.5.1"
sendgrid = "^6.11.0"
motor = "^3.4.0"
pytest = "^8.1.1"
python-multipart = "^0.0.9"
pytest-asyncio = "^0.23.6"
pip = "^24.0"
install = "^1.3.5"

[tool.poetry.dependencies.uvicorn]
extras = [ "standard" ]
version = "^0.23.2"

[tool.poetry.dependencies.pymongo]
version = "^4.6.1"

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"